# Automatically compute the max label width to always properly align the output