@nateberkopec 2025-01-31
Cheers @eregon. Nice article too!